fix: disambiguate the zero-byte addEntry call
uncompressedSize: 0 is ambiguous between addEntry's Int64 overload and
ZIPFoundation's deprecated UInt32 one (Archive+WritingDeprecated.swift:34), and
Swift reports that as a build error rather than picking one. Every other call
site in the suite passes Int64(...) explicitly; this one passed a bare literal.

Caught only by CI, because there is no Swift toolchain on the authoring
machine -- which is the whole reason the branch gates on the build job.

+++ swift/VNPlayerCore/Tests/VNPlayerCoreTests/SaveImporterPlanTests.swift
@@ -391,8 +391,10 @@ final class SaveImporterPlanTests: XCTestCase {
         // one imported save.
         let zip = root.appendingPathComponent("zero-byte.zip")
         let archive = try XCTUnwrap(try? Archive(url: zip, accessMode: .create))
+        // Int64(0), not 0: a bare literal is ambiguous between this overload and
+        // ZIPFoundation's deprecated UInt32 one, and the ambiguity is a build error.
         try archive.addEntry(with: "1-1-LT1.save", type: .file,
-                             uncompressedSize: 0,
+                             uncompressedSize: Int64(0),
                              compressionMethod: .none) { _, _ in Data() }
 
         XCTAssertThrowsError(
